udemy_exercises_intermediate/interfaces/Testability/Testability/OrderProcessor.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Testability
{
    public class OrderProcessor
    {
        private readonly IShippingCalculator _shippingCalculator;//isolating a class using a interface

        public OrderProcessor(IShippingCalculator shippingCalculator)
        {
            _shippingCalculator = shippingCalculator;//loose coupling
        }

        public void Process(Order order)
        {
            if (order.IsShipped)
            
                throw new InvalidOperationException("this order is already processed");

            order.Shipment = new Shipment
            {
                Cost = _shippingCalculator.CalculateShipping(order),
                ShippingDate = DateTime.Today.AddDays(1)
            };
            
        }

    }
}
